V2 release

+++ clients/common/util.c
@@ -8,6 +8,6 @@
 
 void looper(uint32_t delay) __z88dk_fastcall {
     for(uint32_t i = delay; i>0;i--) {
-        zx_border(i%2);
+        zx_border((uint8_t)i%2);
     }
 }
